Hide Service interfaces from users
Many framework Services are not meant to be called by users. By
removing their headers and putting their source code in the plugins
directory it is now impossible for users to access them.
