+++ FWCore/Services/plugins/Timing.cc
@@ -8,8 +8,9 @@
 // Original Author:  Jim Kowalkowski
 //
 
-#include "FWCore/Services/src/Timing.h"
+#include "FWCore/ServiceRegistry/interface/ServiceMaker.h"
 
+#include "FWCore/Utilities/interface/TimingServiceBase.h"
 #include "DataFormats/Provenance/interface/ModuleDescription.h"
 #include "FWCore/MessageLogger/interface/JobReport.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
@@ -27,6 +28,50 @@
 #include <sstream>
 #include <sys/resource.h>
 #include <sys/time.h>
+#include <atomic>
+
+namespace edm {
+  
+  namespace service {
+    class Timing : public TimingServiceBase {
+    public:
+      Timing(ParameterSet const&, ActivityRegistry&);
+      ~Timing();
+      
+      static void fillDescriptions(edm::ConfigurationDescriptions & descriptions);
+      
+      virtual void addToCPUTime(StreamID id, double iTime) override;
+      
+    private:
+      
+      void postBeginJob();
+      void postEndJob();
+      
+      void preEvent(StreamContext const&);
+      void postEvent(StreamContext const&);
+      
+      void preModule(StreamContext const&, ModuleCallingContext const&);
+      void postModule(StreamContext const&, ModuleCallingContext const&);
+      
+      double curr_job_time_;    // seconds
+      double curr_job_cpu_;     // seconds
+                                //use last run time for determining end of processing
+      std::atomic<double> last_run_time_;
+      std::atomic<double> last_run_cpu_;
+      std::vector<double> curr_events_time_;  // seconds
+      bool summary_only_;
+      bool report_summary_;
+      
+      //
+      // Min Max and total event times for each Stream.
+      //  Used for summary at end of job
+      std::vector<double> max_events_time_; // seconds
+      std::vector<double> min_events_time_; // seconds
+      std::vector<double> sum_events_time_;
+      std::atomic<unsigned long> total_event_count_;
+    };
+  }
+}
 
 namespace edm {
   namespace service {
@@ -234,4 +279,9 @@ namespace edm {
     }
   }
 }
+
+using edm::service::Timing;
+
+typedef edm::serviceregistry::AllArgsMaker<edm::TimingServiceBase,Timing> TimingMaker;
+DEFINE_FWK_SERVICE_MAKER(Timing, TimingMaker);
